Split ShapeArray from Arranger. CURA-3239
