+++ resources/qml/Cura.qml
@@ -131,9 +131,9 @@ UM.MainWindow
                 MenuItem { action: Cura.Actions.redo; }
                 MenuSeparator { }
                 MenuItem { action: Cura.Actions.selectAll; }
+                MenuItem { action: Cura.Actions.arrangeAll; }
                 MenuItem { action: Cura.Actions.deleteSelection; }
                 MenuItem { action: Cura.Actions.deleteAll; }
-                MenuItem { action: Cura.Actions.arrangeAll; }
                 MenuItem { action: Cura.Actions.resetAllTranslation; }
                 MenuItem { action: Cura.Actions.resetAll; }
                 MenuSeparator { }
@@ -637,9 +637,9 @@ UM.MainWindow
         MenuItem { action: Cura.Actions.multiplyObject; }
         MenuSeparator { }
         MenuItem { action: Cura.Actions.selectAll; }
+        MenuItem { action: Cura.Actions.arrangeAll; }
         MenuItem { action: Cura.Actions.deleteAll; }
         MenuItem { action: Cura.Actions.reloadAll; }
-        MenuItem { action: Cura.Actions.arrangeSelection; }
         MenuItem { action: Cura.Actions.resetAllTranslation; }
         MenuItem { action: Cura.Actions.resetAll; }
         MenuSeparator { }
@@ -698,9 +698,9 @@ UM.MainWindow
     {
         id: contextMenu;
         MenuItem { action: Cura.Actions.selectAll; }
+        MenuItem { action: Cura.Actions.arrangeAll; }
         MenuItem { action: Cura.Actions.deleteAll; }
         MenuItem { action: Cura.Actions.reloadAll; }
-        MenuItem { action: Cura.Actions.arrangeAll; }
         MenuItem { action: Cura.Actions.resetAllTranslation; }
         MenuItem { action: Cura.Actions.resetAll; }
         MenuSeparator { }
